Fix Sourcery A/I comments

+++ ldap/servers/slapd/auth.c
@@ -396,14 +396,14 @@ handle_bad_certificate(void *clientData, PRFileDesc *prfd)
 
 
 /*
- * Determine if the connection key exchange is Post Quantum Cryptology aware.
+ * Determine if the connection key exchange is Post Quantum Cryptography aware.
  * This function may need to evolve with NSS as more PQC methods get supported.
  */
 static bool
-check_pqc(int connid, SSLChannelInfo *sci)
+check_pqc(uint64_t connid, SSLChannelInfo *sci)
 {
     /*
-     * FYI: To interprets the values:
+     * FYI: To interpret the values:
      * KeaType and KeaGroup values are defined in
      * https://github.com/nss-dev/nss/blob/master/lib/ssl/sslt.h
      * Respectively in SSLKEAType and SSLNamedGroup enums
@@ -411,7 +411,8 @@ check_pqc(int connid, SSLChannelInfo *sci)
     slapi_log_err(SLAPI_LOG_CONNS, "check_pqc", "conn=%" PRIu64 " TLS keaType=%d keaGroup=%d\n",
                   connid, sci->keaType, sci->keaGroup);
 #ifdef MAX_ML_DSA_PRIVATE_KEY_LEN
-    /* PQC KeaType is hybrid */
+    /* NSS supports PQC (because of the ifdef). No let check if the connection use it */
+    /* Check that PQC KeaType is hybrid */
     switch (sci->keaType) {
         case ssl_kea_ecdh_hybrid:
         case ssl_kea_ecdh_hybrid_psk:
@@ -419,7 +420,7 @@ check_pqc(int connid, SSLChannelInfo *sci)
         default:
             return false;
     }
-    /* PQC keaGroup is KEM */
+    /* Check that PQC keaGroup is KEM */
     switch (sci->keaGroup) {
         case ssl_grp_kem_secp256r1mlkem768:
         case ssl_grp_kem_secp384r1mlkem1024:
